Move authorise as function to zendesk folder
